Integrated dual sensor operation with a scaled output
Still needs to have conversion from (what I hope is) inches of water to knots. Other than that, everything is done.

Has a bug: Clicking the buttons can cause the website to crash. Most likely this has something to do with the client requests not being read while the sensors are being polled.